Fix parse bug in `_parse_pad_width` used by `snow2` if boundary_width was a list of lists #bug

+++ porespy/networks/_snow2.py
@@ -244,4 +244,4 @@ def _parse_pad_width(pad_width, shape):
         else:
             raise Exception("pad_width components can't have 2+ elements")
 
-    return np.array(tmp)
+    return np.array(tmp, dtype=int)
